fix: ProdutoResponseDTO error 500 Integer for Long

+++ src/main/java/com/example/integralservice/dto/ProdutoResponseDTO.java
@@ -6,5 +6,5 @@ public record ProdutoResponseDTO(
     String codigo,
     String descricao,
     String unidade,
-    Integer quantidadeAtual
+    Long quantidadeAtual
 ) {}
